requirements.txt
Django==1.9.4
django-debug-toolbar==1.4
djangorestframework==3.3.2
Markdown==2.6.5
mysqlclient==1.3.7
sqlparse==0.1.19
